requirements.txt
fastapi==0.112.2
uvicorn==0.32.1
redis-om==0.3.3
python-decouple==3.8
